Move mockWorld to models to world inside tests
