+++ src/test/java/com/thebois/models/world/MockWorld.java
@@ -1,4 +1,4 @@
-package com.thebois.models;
+package com.thebois.models.world;
 
 import com.thebois.models.world.World;
 import com.thebois.models.world.resources.IResource;
